+++ mobile/src/libs/api/baseApi.ts
@@ -6,6 +6,17 @@ export const baseApi = createApi({
   baseQuery: fetchBaseQuery({
     baseUrl: env.apiBaseUrl,
   }),
-  tagTypes: ['SpotList', 'Exchange', 'ProductTag', 'Watchlist', 'Health'],
+  tagTypes: [
+    'SpotList',
+    'Exchange',
+    'ProductTag',
+    'Watchlist',
+    'Health',
+    'Interval',
+    'Candle',
+    'Ticker',
+    'Supply',
+    'Indicator',
+  ],
   endpoints: () => ({}),
 });